+++ src/side-nav-link/edit.js
@@ -4,13 +4,9 @@ import {
 	InspectorControls,
 	RichText,
 } from '@wordpress/block-editor';
-import {
-	PanelBody,
-	TextControl,
-	SelectControl,
-	ToggleControl,
-} from '@wordpress/components';
+import { PanelBody, SelectControl, ToggleControl } from '@wordpress/components';
 import IconPicker, { iconMaskImage } from '../shared/icon-picker';
+import LinkField from '../shared/link-field';
 
 export default function Edit( { attributes, setAttributes } ) {
 	const { text, href, iconName, isCurrent, matchMode } = attributes;
@@ -23,7 +19,7 @@ export default function Edit( { attributes, setAttributes } ) {
 					title={ __( 'Side nav link', 'awt-blocks' ) }
 					initialOpen={ true }
 				>
-					<TextControl
+					<LinkField
 						label={ __( 'Link URL', 'awt-blocks' ) }
 						value={ href }
 						onChange={ ( value ) =>